Add tests

+++ src/controllers/categoryController.ts
@@ -1,13 +1,17 @@
 import { Router, Request, Response } from 'express';
+
+function escapeRegExp(string: string) {
+	return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+}
 import { ICategory, IController } from '../models/models';
 import { categoryModel, foodModel } from '../models/mongooseSchema';
 import { authAdminToken, authToken } from '../services/tokenService';
 import defaultAnswers from '../helpers/statusCodeHelper';
 import Joi from 'joi';
-import languageBasedErrorMessage from '../helpers/languageHelper';
+import languageBasedMessage from '../helpers/languageHelper';
 import mongoose from 'mongoose';
 
-export default class categoryController implements IController {
+export default class CategoryController implements IController {
 	public router = Router();
 	private category = categoryModel;
 	private food = foodModel;
@@ -18,7 +22,50 @@ export default class categoryController implements IController {
 		this.router.delete('/:id', authAdminToken, this.deleteOne);
 		this.router.put('/:id', authAdminToken, this.modifyOne);
 		this.router.get('', authToken, this.filterCategory);
+		this.router.get('/sub', authToken, this.getSubCategories);
 	}
+
+	private getSubCategories = async (req: Request, res: Response) => {
+		try {
+			const { main } = req.query;
+
+			let subCategories = await this.food.aggregate<ICategory>([
+				{ $match: { categoryId: new mongoose.Types.ObjectId(main as string) } },
+				{ $project: { _id: 0, subCategoryId: 1 } },
+				{ $unwind: '$subCategoryId' },
+				{
+					$lookup: {
+						from: 'categories',
+						localField: 'subCategoryId',
+						foreignField: '_id',
+						as: 'category',
+					},
+				},
+				{ $unwind: '$category' },
+				{
+					$group: {
+						_id: '$category._id',
+						name: { $first: '$category.name' },
+						englishName: { $first: '$category.englishName' },
+						icon: { $first: '$category.icon' },
+					},
+				},
+			]);
+			if (subCategories.length === 0) {
+				throw Error('95');
+			}
+
+			defaultAnswers.ok(
+				res,
+				languageBasedMessage.getLangageBasedName(req, subCategories)
+			);
+		} catch (error: any) {
+			defaultAnswers.badRequest(
+				res,
+				languageBasedMessage.getError(req, error.message)
+			);
+		}
+	};
 
 	private filterCategory = async (req: Request, res: Response) => {
 		try {
@@ -27,13 +74,12 @@ export default class categoryController implements IController {
 				limit = 10,
 				_id,
 				name,
-				englishName,
 				icon,
 				fields,
 				isMainCategory,
 			} = req.query;
 
-			if (isNaN(Number(page)) || isNaN(Number(limit))) {
+			if (Number.isNaN(Number(page)) || Number.isNaN(Number(limit))) {
 				throw Error('93');
 			}
 
@@ -44,9 +90,12 @@ export default class categoryController implements IController {
 
 			const query: any = {};
 			if (_id) query._id = new mongoose.Types.ObjectId(_id as string);
-			if (name) query.name = new RegExp(name as string, 'i');
-			if (englishName)
-				query.englishName = new RegExp(englishName as string, 'i');
+			if (name) {
+				query.$or = [
+					{ name: new RegExp(escapeRegExp(name as string), 'i') },
+					{ englishName: new RegExp(escapeRegExp(name as string), 'i') },
+				];
+			}
 			if (icon) query.icon = new RegExp(icon as string, 'i');
 			if (isMainCategory) query.isMainCategory = isMainCategory === 'true';
 
@@ -72,7 +121,7 @@ export default class categoryController implements IController {
 				};
 			}
 
-			const categories = await this.category.aggregate([
+			const categories: ICategory[] = await this.category.aggregate<ICategory>([
 				{ $match: query },
 				{ $project: projection },
 				{ $skip: skip },
@@ -80,7 +129,7 @@ export default class categoryController implements IController {
 			]);
 
 			res.send({
-				items: categories,
+				items: languageBasedMessage.getLangageBasedName(req, categories),
 				pageCount: Math.ceil(
 					(await this.category.countDocuments(query)) / itemsPerPage
 				),
@@ -88,7 +137,7 @@ export default class categoryController implements IController {
 		} catch (error: any) {
 			defaultAnswers.badRequest(
 				res,
-				languageBasedErrorMessage.getError(req, error.message)
+				languageBasedMessage.getError(req, error.message)
 			);
 		}
 	};
@@ -116,14 +165,14 @@ export default class categoryController implements IController {
 		} catch (error: any) {
 			defaultAnswers.badRequest(
 				res,
-				languageBasedErrorMessage.getError(req, error.message)
+				languageBasedMessage.getError(req, error.message)
 			);
 		}
 	};
 	private deleteOne = async (req: Request, res: Response) => {
 		try {
 			const id = req.params.id;
-			if (id) {
+			if (id && mongoose.Types.ObjectId.isValid(id)) {
 				const response = await this.category.findByIdAndDelete(id);
 				if (response) {
 					defaultAnswers.ok(res, response);
@@ -136,7 +185,7 @@ export default class categoryController implements IController {
 		} catch (error: any) {
 			defaultAnswers.badRequest(
 				res,
-				languageBasedErrorMessage.getError(req, error.message)
+				languageBasedMessage.getError(req, error.message)
 			);
 		}
 	};
@@ -175,7 +224,7 @@ export default class categoryController implements IController {
 		} catch (error: any) {
 			defaultAnswers.badRequest(
 				res,
-				languageBasedErrorMessage.getError(req, error.message)
+				languageBasedMessage.getError(req, error.message)
 			);
 		}
 	};
